+++ Navferty.Common/Extensions/ControlsExtensions.cs
@@ -6,7 +6,7 @@ using System.Windows.Forms;
 
 #nullable enable
 
-namespace NavfertyCommon
+namespace Navferty.Common
 {
 	[DebuggerStepThrough]
 	public static class ControlsExtensions
