+++ Navferty.Common/Extensions/DataExtensions.cs
@@ -7,7 +7,7 @@ using System.Windows.Forms;
 
 #nullable enable
 
-namespace NavfertyCommon
+namespace Navferty.Common
 {
 	[DebuggerStepThrough]
 	public static class DataExtensions
