+++ Navferty.Common/Interfaces/IDialogService.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace NavfertyCommon
+namespace Navferty.Common
 {
 	public interface IDialogService
 	{
