+++ Navferty.Common/Properties/AssemblyInfo.cs
