+++ NavfertyExcelAddIn/ParseNumerics/DecimalParser.cs
@@ -1,102 +1,102 @@
 ﻿using System.Globalization;
 using System.Text.RegularExpressions;
 
-using NavfertyCommon;
+using Navferty.Common;
 
 namespace NavfertyExcelAddIn.ParseNumerics
 {
-	/// <summary>
-	/// Parser implemented in https://github.com/navferty/NumericParser
-	/// </summary>
-	public static class DecimalParser
-	{
-		private static readonly Regex SpacesPattern = new(@"\s");
-		private static readonly Regex DecimalPattern = new(@"[\d\.\,\s]*");
-		private static readonly Regex ExponentPattern = new(@"[-+]?\d*\.?\d+[eE][-+]?\d+");
+    /// <summary>
+    /// Parser implemented in https://github.com/navferty/NumericParser
+    /// </summary>
+    public static class DecimalParser
+    {
+        private static readonly Regex SpacesPattern = new(@"\s");
+        private static readonly Regex DecimalPattern = new(@"[\d\.\,\s]*");
+        private static readonly Regex ExponentPattern = new(@"[-+]?\d*\.?\d+[eE][-+]?\d+");
 
-		//private const string CS10_TEST = $"{{222";
-		private static readonly string? fff = null;
+        //private const string CS10_TEST = $"{{222";
+        private static readonly string? fff = null;
 
-		public static decimal? ParseDecimal(this string value)
-		{
-			if (string.IsNullOrWhiteSpace(value))
-			{
-				return null;
-			}
+        public static decimal? ParseDecimal(this string value)
+        {
+            if (string.IsNullOrWhiteSpace(value))
+            {
+                return null;
+            }
 
-			var v = SpacesPattern.Replace(value, match => string.Empty);
+            var v = SpacesPattern.Replace(value, match => string.Empty);
 
-			if (ExponentPattern.IsMatch(v))
-			{
-				return v.TryParseExponent();
-			}
+            if (ExponentPattern.IsMatch(v))
+            {
+                return v.TryParseExponent();
+            }
 
-			if (!DecimalPattern.IsMatch(value))
-			{
-				return null;
-			}
+            if (!DecimalPattern.IsMatch(value))
+            {
+                return null;
+            }
 
-			if (v.Contains(",") && v.Contains("."))
-			{
-				var last = v.LastIndexOfAny(new[] { ',', '.' });
-				var c = v[last];
-				return v.CountChars(c) == 1
-					? v.TryParse(c == '.' ? Format.Dot : Format.Comma)
-					: null;
-			}
+            if (v.Contains(",") && v.Contains("."))
+            {
+                var last = v.LastIndexOfAny(new[] { ',', '.' });
+                var c = v[last];
+                return v.CountChars(c) == 1
+                    ? v.TryParse(c == '.' ? Format.Dot : Format.Comma)
+                    : null;
+            }
 
-			if (v.Contains(","))
-			{
-				return v.CountChars(',') == 1
-					? v.TryParse(Format.Comma)
-					: v.TryParse(Format.Dot);
-			}
+            if (v.Contains(","))
+            {
+                return v.CountChars(',') == 1
+                    ? v.TryParse(Format.Comma)
+                    : v.TryParse(Format.Dot);
+            }
 
-			if (v.Contains("."))
-			{
-				return v.CountChars('.') == 1
-					? v.TryParse(Format.Dot)
-					: v.TryParse(Format.Comma);
-			}
+            if (v.Contains("."))
+            {
+                return v.CountChars('.') == 1
+                    ? v.TryParse(Format.Dot)
+                    : v.TryParse(Format.Comma);
+            }
 
-			return v.TryParse(Format.Dot);
-		}
+            return v.TryParse(Format.Dot);
+        }
 
-		private static decimal? TryParseExponent(this string value)
-		{
-			return decimal.TryParse(value, NumberStyles.Float, CultureInfo.InvariantCulture, out decimal result)
-				? result
-				: (decimal?)null;
-		}
+        private static decimal? TryParseExponent(this string value)
+        {
+            return decimal.TryParse(value, NumberStyles.Float, CultureInfo.InvariantCulture, out decimal result)
+                ? result
+                : (decimal?)null;
+        }
 
-		private static decimal? TryParse(this string value, Format info)
-		{
-			var formatInfo = (NumberFormatInfo)NumberFormatInfo.InvariantInfo.Clone();
+        private static decimal? TryParse(this string value, Format info)
+        {
+            var formatInfo = (NumberFormatInfo)NumberFormatInfo.InvariantInfo.Clone();
 
-			if (info == Format.Comma)
-			{
-				formatInfo.CurrencyDecimalSeparator = ",";
-				formatInfo.CurrencyGroupSeparator = ".";
-				formatInfo.NumberDecimalSeparator = ",";
-				formatInfo.NumberGroupSeparator = ".";
-			}
-			else
-			{
-				formatInfo.CurrencyDecimalSeparator = ".";
-				formatInfo.CurrencyGroupSeparator = ",";
-			}
-			// добавить тест-кейсов на формат валют
-			//formatInfo.CurrencyNegativePattern = 8;
-			//formatInfo.CurrencyPositivePattern = 3;
-			return decimal.TryParse(value, NumberStyles.Currency, formatInfo, out decimal result)
-				? result
-				: (decimal?)null;
-		}
+            if (info == Format.Comma)
+            {
+                formatInfo.CurrencyDecimalSeparator = ",";
+                formatInfo.CurrencyGroupSeparator = ".";
+                formatInfo.NumberDecimalSeparator = ",";
+                formatInfo.NumberGroupSeparator = ".";
+            }
+            else
+            {
+                formatInfo.CurrencyDecimalSeparator = ".";
+                formatInfo.CurrencyGroupSeparator = ",";
+            }
+            // добавить тест-кейсов на формат валют
+            //formatInfo.CurrencyNegativePattern = 8;
+            //formatInfo.CurrencyPositivePattern = 3;
+            return decimal.TryParse(value, NumberStyles.Currency, formatInfo, out decimal result)
+                ? result
+                : (decimal?)null;
+        }
 
-		private enum Format
-		{
-			Dot,
-			Comma
-		}
-	}
+        private enum Format
+        {
+            Dot,
+            Comma
+        }
+    }
 }
