+++ NavfertyExcelAddIn/Web/CurrencyExchangeRates/CurrencyExchangeRatesBuilder.cs
@@ -1,7 +1,7 @@
 ﻿
 using Microsoft.Office.Interop.Excel;
 
-using NavfertyCommon;
+using Navferty.Common;
 
 using NavfertyExcelAddIn.Localization;
 
@@ -9,29 +9,29 @@ using NavfertyExcelAddIn.Localization;
 
 namespace NavfertyExcelAddIn.Web.CurrencyExchangeRates
 {
-	public class CurrencyExchangeRatesBuilder : ICurrencyExchangeRates
-	{
-		internal readonly IDialogService dialogService;
-		private Microsoft.Office.Interop.Excel.Application App => Globals.ThisAddIn.Application;
+    public class CurrencyExchangeRatesBuilder : ICurrencyExchangeRates
+    {
+        internal readonly IDialogService dialogService;
+        private Microsoft.Office.Interop.Excel.Application App => Globals.ThisAddIn.Application;
 
-		public CurrencyExchangeRatesBuilder(IDialogService dialogService)
-			=> this.dialogService = dialogService;
+        public CurrencyExchangeRatesBuilder(IDialogService dialogService)
+            => this.dialogService = dialogService;
 
 
-		public void ShowCurrencyExchangeRates(Workbook wb)
-		{
-			Range? sel = App.Selection;
-			if (null == sel || sel.Cells == null || sel.Cells.Count < 1)
-			{
-				dialogService.ShowError(UIStrings.CurrencyExchangeRates_Error_NedAnyCellSelection);
-				return;
-			}
+        public void ShowCurrencyExchangeRates(Workbook wb)
+        {
+            Range? sel = App.Selection;
+            if (null == sel || sel.Cells == null || sel.Cells.Count < 1)
+            {
+                dialogService.ShowError(UIStrings.CurrencyExchangeRates_Error_NedAnyCellSelection);
+                return;
+            }
 
-			var rslt = frmExchangeRates.SelectExchageRate(dialogService);
-			if (rslt == null) return;//User cancel
+            var rslt = frmExchangeRates.SelectExchageRate(dialogService);
+            if (rslt == null) return;//User cancel
 
-			var exchangeRate = rslt.CursFor1Unit;
-			sel.Value = exchangeRate;
-		}
-	}
+            var exchangeRate = rslt.CursFor1Unit;
+            sel.Value = exchangeRate;
+        }
+    }
 }
