+++ NavfertyExcelAddIn/WorksheetCellsEditing/EmptySpaceTrimmer.cs
@@ -2,7 +2,7 @@
 
 using Microsoft.Office.Interop.Excel;
 
-using NavfertyCommon;
+using Navferty.Common;
 
 using NavfertyExcelAddIn.Commons;
 
@@ -10,28 +10,28 @@ using NLog;
 
 namespace NavfertyExcelAddIn.WorksheetCellsEditing
 {
-	public class EmptySpaceTrimmer : IEmptySpaceTrimmer
-	{
-		private static readonly ILogger logger = LogManager.GetCurrentClassLogger();
+    public class EmptySpaceTrimmer : IEmptySpaceTrimmer
+    {
+        private static readonly ILogger logger = LogManager.GetCurrentClassLogger();
 
-		public void TrimExtraSpaces(Range range)
-		{
-			logger.Info($"Trim spaces for range {range.GetRelativeAddress()}");
+        public void TrimExtraSpaces(Range range)
+        {
+            logger.Info($"Trim spaces for range {range.GetRelativeAddress()}");
 
-			range.ApplyForEachCellOfType<string, string>(value => value.TrimSpaces());
-		}
+            range.ApplyForEachCellOfType<string, string>(value => value.TrimSpaces());
+        }
 
-		public void RemoveAllSpaces(Range range)
-		{
-			logger.Info($"Trim spaces for range {range.GetRelativeAddress()}");
+        public void RemoveAllSpaces(Range range)
+        {
+            logger.Info($"Trim spaces for range {range.GetRelativeAddress()}");
 
-			range.ApplyForEachCellOfType<string, string>(value =>
-			{
-				if (string.IsNullOrWhiteSpace(value))
-					return null;
+            range.ApplyForEachCellOfType<string, string>(value =>
+            {
+                if (string.IsNullOrWhiteSpace(value))
+                    return null;
 
-				return new string(value.Where(c => !char.IsWhiteSpace(c)).ToArray());
-			});
-		}
-	}
+                return new string(value.Where(c => !char.IsWhiteSpace(c)).ToArray());
+            });
+        }
+    }
 }
